Commit thoughts before bed
My amount of paths to each splitter isn't propagating the way I want it
to. And there's some obvious bug that I', nopt seing at the moment
because I'm tired.

it should be like S -> 1
then the 1 propagates downward, hits a splitter, then those paths each
propagate a 1 downward, then we hit the important splitter that has an
overlap:

...S...
...1...
..1^1..
..1.1..
..^.^..
.1.2.1.

That 2 should come from adding up the paths from the split each. Which
should be as simple as += 1 to the same place as we split, but we dont
for some reason