D9P2 - complete - geometry solution
Someone chatting in IRC shared a haskell solution here [1] and it
was interesting, and made me go

	OH NO ITS DAY 13 2024 ALL OVER AGAIN

Attempting solutions that sound like they should maybe work but don't
because of the intractable search space. I did flood fill, ran it for
30m and then watched it use up waaay too much ram. I did the odd-even
test for every point in the rectangle, that is ALSO too time intensive,
there are simply too many points!

So stumped. I looked for hints like seeing folks chatter about it and
then finding the: ah fuck it's geometry again. situation. I'm not
particularly GOOD at geometry (see my bounding box issues on the blag)
and so it doesn't surprise me that I needed a lot of help with this one.

Even with the haskell sample to go off of though, it was still hard to
find the neccesary learning materials, and so I had gippity tutor me a
bit. Or at least tried to. It's really really bad at NOT giving the
answer a lot of the time which is both frustating and helpful.

[1]:https://github.com/liquidcake1/advent-of-code/commit/ce3f7ab253ab46e4be58fc1551252a641fc08429