D12P1 area method
I'm thinking that maybe I can rule out a whole bunch by taking the total
area of the shapes requested for the region, and then comparing it to
the region itself.

If the region can be optimally packed, then the area of the shapes will
be less than the total region after all.